Add examenes y retroalimentacion

+++ database/migrations/2024_03_25_225147_create_calificacions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('calificacions', function (Blueprint $table) {
             $table->id();
             $table->string('comentario')->nullable();
-            $table->integer('nota')->nullable();
+            $table->float('nota')->nullable();
             $table->boolean('finalizado')->nullable();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('ejecucion_id');
